# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DNS Markdown API",
    "version": "1.0.0",
    "description": "Generated from zod request schemas. See API.md for auth/CORS/error format."
  },
  "components": {
    "securitySchemes": {
      "clerkAndApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-API-Secret",
        "description": "Requires an active Clerk session cookie plus this shared secret header."
      },
      "serviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal service-to-service token (API_SERVICE_KEY)."
      },
      "svixWebhook": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-signature",
        "description": "Verified via svix; not usable from Swagger UI's Try it out."
      }
    }
  },
  "paths": {
    "/api/pricelist": {
      "get": {
        "summary": "Get the latest price list for a city",
        "tags": ["Pricelist"],
        "security": [
          {
            "clerkAndApiSecret": []
          }
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/user": {
      "post": {
        "summary": "Get the current user's profile",
        "tags": ["User"],
        "security": [
          {
            "clerkAndApiSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/clerk/create-user": {
      "post": {
        "summary": "Clerk user.created webhook (svix-signed)",
        "tags": ["Clerk"],
        "security": [
          {
            "svixWebhook": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    }
  }
}